+++ CHANGELOG.md
@@ -24,6 +24,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * `jj desc` adds the commit trailers, configured in the `commit_trailers` template, to the commit description. Use
   cases include DCO Sign Off and Gerrit Change Id.
 
+* `jj commit` now uses the `commit_trailers` template.
+
 ### Fixed bugs
 
 ## [0.28.0] - 2025-04-02